refactor(tunnel): split pending UID UDP storage
